2023-12-24 20:46 UTC-0500 Ron Pinkas <ronpinkas/AT/gmail/com>   * include/hbapi.h   * source/compiler/comptool.c   * source/vm/cmdarg.c   * source/pp/ppgen.c     * Corrected hb_verCvsID() to return ULONGLONG instead of long long int   * source/pp/ppgen.c     Corrected sufix of 64bit version consttants from LL to UL

+++ source/compiler/comptool.c
@@ -626,7 +626,7 @@ int hb_setGetDirSeparator( void )
 }
 
 /* ChangeLog CVS revision number */
-long long int hb_verCvsID( void )
+ULONGLONG hb_verCvsID( void )
 {
    return HB_VER_CVSID;
 }
