+++ source/vm/cmdarg.c
@@ -411,7 +411,7 @@ void hb_cmdargProcessVM( void )
 }
 
 /* ChangeLog CVS revision number */
-long long int hb_verCvsID( void )
+ULONGLONG hb_verCvsID( void )
 {
    return HB_VER_CVSID;
 }
